requirements.txt
pip>=23.2
setuptools>=67.0.0
wheel>=0.38.0
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
Pillow==10.0.0
ultralytics==8.0.181